Split HistoryTreeProcessor in views and service
